[breadboard-web] Unify loading
